+++ src/divider/index.tsx
@@ -1,2 +1,0 @@
-export { Divider, DividerProps } from './Divider'
-export { Divider as default } from './Divider'