+++ src/divider/sectionDivider/SectionDivider.unit.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import { render, screen } from '@testing-library/react'
+
+import { SectionDivider } from '.'
+
+describe('SectionDivider', () => {
+  it('Should have a separator role', () => {
+    render(<SectionDivider />)
+    expect(screen.getByRole('separator')).toBeInTheDocument()
+  })
+})