+++ src/layout/section/highlightSection/story.tsx
@@ -2,7 +2,7 @@ import React, { Fragment } from 'react'
 import { withKnobs } from '@storybook/addon-knobs'
 import { storiesOf } from '@storybook/react'
 
-import { BlankSeparator } from '../../../blankSeparator'
+import { SpacingDivider } from '../../../divider/spacingDivider'
 import { HighlightSection } from './index'
 
 const stories = storiesOf('Sections|HighlightSection', module)
@@ -10,7 +10,7 @@ stories.addDecorator(withKnobs)
 
 stories.add('default', () => (
   <Fragment>
-    <BlankSeparator />
+    <SpacingDivider />
     <HighlightSection title="Section title">
       This some content for the highlighted section!
       <br />
@@ -18,6 +18,6 @@ stories.add('default', () => (
       <br />
       Some more more content.
     </HighlightSection>
-    <BlankSeparator />
+    <SpacingDivider />
   </Fragment>
 ))